Fix type-conversion warnings from MSVC

+++ language/editor/modificationrevisionset.cpp
@@ -68,7 +68,7 @@ struct FileModificationPairRequest {
     return m_data.hash();
   }
 
-  size_t itemSize() const {
+  uint itemSize() const {
       return m_data.itemSize();
   }
