+++ language/util/basicsetrepository.h
@@ -149,7 +149,7 @@ public:
   }
   
   //Should return the size of an item created with createItem
-  inline size_t itemSize() const {
+  inline uint itemSize() const {
       return sizeof(SetNodeData);
   }
   //Should create an item where the information of the requested item is permanently stored. The pointer
